Fix integer overflow issue
With the previous code, if the first level had no max target, while the second did, it would try to add the second's max target to `int.MaxValue`, which would cause an integer overflow and mess up the min/max logic in the `Aggregate`s.

Now, the code never overflows so the logic makes sense again.